refactored and cleans up github login
