+++ e2e/studio/package.json
@@ -16,8 +16,7 @@
     "@supabase/supabase-js": "catalog:",
     "cross-fetch": "^4.1.0",
     "dotenv": "^16.5.0",
-    "otpauth": "^9.4.1",
-    "playwright": "^1.57.0"
+    "otpauth": "^9.4.1"
   },
   "devDependencies": {
     "@faker-js/faker": "^9.9.0",
